created funtioning outfit list with moch data, created file for client side helper functions, cleaned up code
